+++ src/main/java/com/nisovin/shopkeepers/tradelog/data/TradeRecord.java
@@ -6,11 +6,36 @@ import org.bukkit.inventory.ItemStack;
 
 import com.nisovin.shopkeepers.api.events.ShopkeeperTradeEvent;
 import com.nisovin.shopkeepers.util.Validate;
+import com.nisovin.shopkeepers.util.trading.MergedTrades;
 
 /**
- * An immutable snapshot of the information about a trade that happened.
+ * An immutable snapshot of the information about one or multiple equal trades that took place.
+ * <p>
+ * In order to represent trades more compactly, a single {@link TradeRecord} may represent a number of consecutive
+ * trades that involve the same player, the same shopkeeper, and the same traded items.
  */
 public class TradeRecord {
+
+	/**
+	 * Creates a {@link TradeRecord} for the given {@link MergedTrades}.
+	 * 
+	 * @param trades
+	 *            the merged trades
+	 * @return the trade record
+	 */
+	public static TradeRecord create(MergedTrades trades) {
+		Validate.notNull(trades, "trades is null");
+		Instant timestamp = trades.getTimestamp();
+		ShopkeeperTradeEvent tradeEvent = trades.getInitialTrade();
+		PlayerRecord playerRecord = PlayerRecord.of(tradeEvent.getPlayer());
+		ShopRecord shopRecord = ShopRecord.of(tradeEvent.getShopkeeper());
+		// We reuse the items of the given MergedTrades:
+		ItemStack item1 = trades.getOfferedItem1();
+		ItemStack item2 = trades.getOfferedItem2();
+		ItemStack resultItem = trades.getResultItem();
+		int tradeCount = trades.getTradeCount();
+		return new TradeRecord(timestamp, playerRecord, shopRecord, item1, item2, resultItem, tradeCount);
+	}
 
 	/**
 	 * Creates a {@link TradeRecord} for the given {@link ShopkeeperTradeEvent}.
@@ -25,35 +50,38 @@ public class TradeRecord {
 		PlayerRecord playerRecord = PlayerRecord.of(tradeEvent.getPlayer());
 		ShopRecord shopRecord = ShopRecord.of(tradeEvent.getShopkeeper());
 		// These ItemStacks are clones:
-		ItemStack resultItem = tradeEvent.getTradingRecipe().getResultItem();
 		ItemStack item1 = tradeEvent.getOfferedItem1();
 		ItemStack item2 = tradeEvent.getOfferedItem2();
-		return new TradeRecord(timestamp, playerRecord, shopRecord, resultItem, item1, item2);
+		ItemStack resultItem = tradeEvent.getTradingRecipe().getResultItem();
+		return new TradeRecord(timestamp, playerRecord, shopRecord, item1, item2, resultItem, 1);
 	}
 
 	private final Instant timestamp; // Not null
 	private final PlayerRecord player; // Not null
 	private final ShopRecord shop; // Not null
-	private final ItemStack resultItem; // Not null
 	// The items provided by the player that match the first and second items required by the trade. These items might
 	// not necessarily be equal the items required by the trade (they only have to 'match' / be accepted). Their amounts
 	// match those of the trading recipe.
 	// The order in which the player provided the items in the trading interface is not recorded.
 	private final ItemStack item1; // Not null
 	private final ItemStack item2; // Can be null
+	private final ItemStack resultItem; // Not null
+	private final int tradeCount;
 
-	private TradeRecord(Instant timestamp, PlayerRecord player, ShopRecord shop, ItemStack resultItem, ItemStack item1, ItemStack item2) {
+	private TradeRecord(Instant timestamp, PlayerRecord player, ShopRecord shop, ItemStack item1, ItemStack item2, ItemStack resultItem, int tradeCount) {
 		Validate.notNull(timestamp, "timestamp is null");
 		Validate.notNull(player, "player is null");
 		Validate.notNull(shop, "shop is null");
-		Validate.notNull(resultItem, "resultItem is null");
 		Validate.notNull(item1, "item1 is null");
+		Validate.notNull(resultItem, "resultItem is null");
+		Validate.isTrue(tradeCount > 0, "tradeCount has to be positive");
 		this.timestamp = timestamp;
 		this.player = player;
 		this.shop = shop;
-		this.resultItem = resultItem;
 		this.item1 = item1;
 		this.item2 = item2;
+		this.resultItem = resultItem;
+		this.tradeCount = tradeCount;
 	}
 
 	/**
@@ -81,18 +109,6 @@ public class TradeRecord {
 	 */
 	public ShopRecord getShop() {
 		return shop;
-	}
-
-	/**
-	 * Gets the result item.
-	 * <p>
-	 * In order to avoid excessive item copying, this returns the item stored by this record without copying it first.
-	 * However, it is only meant for read-only purposes. Do not modify it!
-	 * 
-	 * @return the result item, not <code>null</code>
-	 */
-	public ItemStack getResultItem() {
-		return resultItem;
 	}
 
 	/**
@@ -125,6 +141,27 @@ public class TradeRecord {
 		return item2;
 	}
 
+	/**
+	 * Gets the result item.
+	 * <p>
+	 * In order to avoid excessive item copying, this returns the item stored by this record without copying it first.
+	 * However, it is only meant for read-only purposes. Do not modify it!
+	 * 
+	 * @return the result item, not <code>null</code>
+	 */
+	public ItemStack getResultItem() {
+		return resultItem;
+	}
+
+	/**
+	 * Gets the number of equivalent trades that are represented by this {@link TradeRecord}.
+	 * 
+	 * @return the number of trades
+	 */
+	public int getTradeCount() {
+		return tradeCount;
+	}
+
 	@Override
 	public String toString() {
 		StringBuilder builder = new StringBuilder();
@@ -134,12 +171,14 @@ public class TradeRecord {
 		builder.append(player);
 		builder.append(", shop=");
 		builder.append(shop);
-		builder.append(", resultItem=");
-		builder.append(resultItem);
 		builder.append(", item1=");
 		builder.append(item1);
 		builder.append(", item2=");
 		builder.append(item2);
+		builder.append(", resultItem=");
+		builder.append(resultItem);
+		builder.append(", tradeCount=");
+		builder.append(tradeCount);
 		builder.append("]");
 		return builder.toString();
 	}
@@ -151,26 +190,27 @@ public class TradeRecord {
 		result = prime * result + timestamp.hashCode();
 		result = prime * result + player.hashCode();
 		result = prime * result + shop.hashCode();
-		result = prime * result + resultItem.hashCode();
 		result = prime * result + item1.hashCode();
 		result = prime * result + ((item2 == null) ? 0 : item2.hashCode());
+		result = prime * result + resultItem.hashCode();
+		result = prime * result + tradeCount;
 		return result;
 	}
 
 	@Override
 	public boolean equals(Object obj) {
 		if (this == obj) return true;
-		if (obj == null) return false;
 		if (!(obj instanceof TradeRecord)) return false;
 		TradeRecord other = (TradeRecord) obj;
 		if (!timestamp.equals(other.timestamp)) return false;
+		if (tradeCount != other.tradeCount) return false;
 		if (!player.equals(other.player)) return false;
 		if (!shop.equals(other.shop)) return false;
-		if (!resultItem.equals(other.resultItem)) return false;
 		if (!item1.equals(other.item1)) return false;
 		if (item2 == null) {
 			if (other.item2 != null) return false;
 		} else if (!item2.equals(other.item2)) return false;
+		if (!resultItem.equals(other.resultItem)) return false;
 		return true;
 	}
 }